DX-1071: Rename jekyll_environment*.rb to exec_env*.rb
